refactor: move component folders
