+++ src/app/components/MainCategorySection.tsx
